Define postprocessing api and move face processing to fit
